requirements.txt
streamlit>=1.40,<2.0
PyYAML>=6.0,<7.0

